CAMEL-16400: split unit and integration tests (azure components)
* split unit and integration tests for camel-azure-eventhubs
* split unit and integration tests for camel-azure-storage-blob
* split unit and integration tests for camel-azure-storage-datalake
* split unit and integration tests for camel-azure-storage-queue
