+++ components/camel-azure/camel-azure-storage-blob/src/test/java/org/apache/camel/component/azure/storage/blob/integration/BlobContainerOperationsIT.java
@@ -36,7 +36,7 @@ import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-class BlobContainerOperationsITTest extends BaseIT {
+class BlobContainerOperationsIT extends Base {
 
     private BlobServiceClientWrapper blobServiceClientWrapper;
