+++ components/camel-azure/camel-azure-storage-datalake/src/test/java/org/apache/camel/component/azure/storage/datalake/integration/Base.java
@@ -33,17 +33,16 @@ import org.apache.camel.test.infra.azure.storage.datalake.services.AzureStorageD
 import org.apache.camel.test.junit5.CamelTestSupport;
 import org.apache.commons.lang3.RandomStringUtils;
 import org.apache.commons.lang3.StringUtils;
-import org.junit.Ignore;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.TestInstance;
 import org.junit.jupiter.api.extension.RegisterExtension;
 
-@Ignore
+//@EnabledIfSystemProperty(named = "azure.instance.type", matches = "remote")
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
-public class BaseIT extends CamelTestSupport {
+public class Base extends CamelTestSupport {
 
     @RegisterExtension
-    public static AzureService service;
+    public AzureService service = AzureStorageDataLakeServiceFactory.createService();
 
     protected DataLakeServiceClient serviceClient;
     protected DataLakeConfiguration configuration;
@@ -51,7 +50,6 @@ public class BaseIT extends CamelTestSupport {
 
     static {
         initCredentials();
-        service = AzureStorageDataLakeServiceFactory.createService();
     }
 
     @Override
@@ -68,7 +66,6 @@ public class BaseIT extends CamelTestSupport {
         if (StringUtils.isNotEmpty(accountName) && StringUtils.isNotEmpty(accessKey)) {
             System.setProperty(AzureConfigs.ACCOUNT_NAME, accountName);
             System.setProperty(AzureConfigs.ACCOUNT_KEY, accessKey);
-            System.setProperty("azure.instance.type", "remote");
         }
     }
