+++ components/camel-azure/camel-azure-storage-datalake/src/test/java/org/apache/camel/component/azure/storage/datalake/integration/DataLakeFileOperationIT.java
@@ -38,16 +38,16 @@ import org.apache.camel.component.azure.storage.datalake.operations.DataLakeOper
 import org.apache.camel.support.DefaultExchange;
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.lang3.RandomStringUtils;
-import org.junit.Ignore;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.condition.EnabledIfSystemProperty;
 import org.junit.jupiter.api.io.TempDir;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-@Ignore
-public class DataLakeFileOperationIT extends BaseIT {
+@EnabledIfSystemProperty(named = "azure.instance.type", matches = "remote")
+public class DataLakeFileOperationIT extends Base {
     private DataLakeFileSystemClientWrapper fileSystemClientWrapper;
     private String randomFileName;
